+++ packages/browseros-agent/apps/server/src/api/services/terminal/terminal-protocol.ts
